backend_server/requirements.txt
Django==5.2.6
psycopg[binary]==3.2.10
python-dotenv==1.1.1
dj-database-url==3.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0
Pillow==10.4.0
PyJWT==2.9.0
drf-spectacular==0.27.2
